requirements.txt
certifi==2017.4.17
chardet==3.0.4
Django==1.10.6
django-crispy-forms==1.6.1
django-oauth-toolkit==1.0.0
djangorestframework==3.6.3
gunicorn==19.7.1
idna==2.5
oauthlib==2.0.2
python-decouple==3.0
requests==2.18.1
urllib3==1.21.1
virtualenv==15.1.0
whitenoise==3.3.0
